rename muncher file name and remove duplicated files
